LPS-124483 Move LifecycleListener and remove files and code that is not needed. We now call the site initializer, creating the home page
